Enhance README with project details and analysis
This project utilizes R for descriptive statistical analysis of patterns in the generation of sunspots, including basic data visualization. The raw dataset used for analysis contains centuries of observations of sunspots. Sunspots are a product of the solar cycle with a documented periodicity of 11 years. We utilized a t-test and produced a 95% confidence interval to confirm that that the mean length of a solar cycle is approximately 11 years.